PWS/requirements.txt
aiohttp
aiofiles
aio-request
pycryptodome

